requirements.txt
Babel==2.10.3
black==22.8.0
Brotli==1.0.9
click==8.1.3
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.12
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.2
pandas==1.4.4
pandas-stubs==1.4.4.220906
pathspec==0.10.1
platformdirs==2.5.2
plotly==5.10.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
python-i18n==0.3.9
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tenacity==8.0.1
tomli==2.0.1
types-pytz==2022.2.1.0
Werkzeug==2.2.2
